Fix various bugs
  * Fixed nic flush on each iteration. If more than one network
was configured for a nic, that nics IP and routes would be flushed
on each iteration. The result being that only the last IP was set
  * Fixed conflicting routes when multiple networks were configured.
You can only have one route to a particular destination with a
particular metric. This simply skips any conflicting route, but future
iterations should maybe set a different metric.
  * Fixed duplicate nameservers in netplan config